Make all SystemUIDialog's the same configurable width (1/2)
This CL sets the default width of SystemUIDialog to be MATCH_PARENT on
phones and 624dp on tablets/foldables.

Moreover, as part of b/203389579, the width on tablets/foldables can be
configured using a system property. That configurable width will be
removed once b/203389579 is fixed, before the sc-v2 release.

Bug: 203389579
Test: Manual
Change-Id: I5a54914666415c9aa278d42492d9339480222df0

Former-commit-id: 4de63945eab26b78591e776acec98d7f0dca52c2