Respect the expanded amount with AUTH_SCRIMMED_SHADE
ScrimState.UNLOCKED is used when the device is locked
and an activity is occluding the keyguard (ie: camera activty
on top of the lock screen, then pull down the shade). Therefore,
the AUTH_SCRIMMED_SHADE (tap notification with udfps enrolled) needs to
also respect the expanded state of the shade to correctly
determine the behind and notification scrim alpha values.

Test: atest ScrimControllerTest
Fixes: 198789984
Change-Id: I36f064c84821389367a06edff736bed8a75cdc19

Former-commit-id: e764ba30d79178e1a9ad6fbd8bf12ee245a40983