PiP: Fix more Shell Transition potential issues.
If an animator is already running, we usually just update its end
bounds - but we also end up over-writing the original callbacks. This
used to work well since pre-shell transition, callbacks are always
pulled from PipTaskOrganizer, but now we can have two places that supply
callbacks (PipTransition). So let's just leave the callback alone.

Another potential issue is that sometimes finish callbacks from one
transition is not called before starting another one, and starting a new
transition overwrites the old finishCallback. We will then finish the
earlier callbacks to ensure the transition system does not hang the
entire device, but throw an Exception to not gloss over the potential
issue.

Bug: 183993924
Test: atest PinnedStackTests
Change-Id: Ib4c5926beec5745d8a0abccec5c67849488e532b

Former-commit-id: f71a1b2970c6a2c93201d38d4053814d3f3be392