Display's transition phase checked before DPC init.
DisplayPowerController checks to see if a display is in a transition as
part of it's display state calculation. However, this was happening only
after the initialization call in DPC on the first loop of
updatePowerState.  This meant that a display added in a TRANSITION phase
would get incorrectly initialized to ON instead of OFF.

Test: Manually verify in logs that new displays are not initialized to ON
during the fold action of a foldable device.
Bug: 199123716

Change-Id: I8a2b3154cf4c36bc5fe924e787a3db1d11933f79

Former-commit-id: 63c9e5fa4a6afa00ef531ab2d022f16326257c0e