Add RemoteInputConnectionImpl#finishComposingTextFromImm()
This is a preparation CL to make InputConnection tasks cancellable.

Currently

  RemoteInputConnectionImpl#finishComposingText()

is the only method that can be called by both RemoteInputConnection as
a Binder call and InputMethodManager as a direct invocation.  This is
not convenient because a common header data structure is going to be
added to every IPC defined in IInputContext in order to implement
InputConnection task cancelation.

Thus, this CL introduces

  RemoteInputConnectionImpl#finishComposingTextFromImm()

to decouple this code path from subsequent CLs.

There should be no observable behavior change yet in this CL.

Bug: 195115071
Test: presubmit
Change-Id: I07290af2595c7a556cb5daa8a4c0f9c4a7daa49c

Former-commit-id: e43a10f257a109b533326bf251ecc167a47fe769