Move all the window layout logic to WindowLayout
This CL removes redundant layout logic in computeFrame and moves all the
window layout logic to WindowLayout. We don't need mContainingFrame in
WindowFrames anymore.

The goal in the future is to let the client side to invoke
WindowLayout#computeWindowFrames, and then send the frames to the
server. The server will only use the method to layout the insets source
windows.

Bug: 161810301
Bug: 175858823
Test: atest DisplayPolicyLayoutTests WindowLayoutTests
            DialogFrameTests ManifestLayoutTests ToastWindowTest
Change-Id: I91aa8051970cb8f30f4c13849a4db79b16182075

Former-commit-id: 5210f9646fa739e898f8c44ee8274387d1acb699