Add sensitive content redaction to notif pipeline
This incorporates two behaviors originally handled by the
NotificationViewHierarchyManager:

1. A change in DynamicPrivacy would trigger
   NotificationViewHierarchyManager#updateNotificationViews

2. The aformentioned #updateNotificationViews method was responsible
   for setting the redaction state on each NotificationEntry

A new Coordinator is introduced to the NotifPipeline to handle both of
these behaviors. (1) is handled by attaching a new Invalidator that
will request a notif list rebuild when dynamic privacy changes. (2) is
handled by processing each ListEntry in an OnPreRenderListListener,
and setting the redaction state on each NotificationEntry.

Test: manual - turn on setting to hide sensitive notification content
               on the lockscreen, lock phone, turn screen back on,
               observe notifications on lock screen
Fixes: 200025730
Change-Id: I57b4d68f440ea50f13f6da4676144400f875ea8c

Former-commit-id: 6a9124a0a54adede492466adbc8a2da5146c4fd8