Add required flag to registerReceiver call in WindowMagnificationPC
Android T adds support to allow a runtime receiver to be registered as
not exported, but to ensure apps can take advantage of this, calls to
registerReceiver must specify a flag indicating whether the receiver
should be exported for apps targeting T+ that are registering for
unprotected broadcasts. This commit adds the RECEIVER_EXPORTED flag
to the call to registerReceiver in WindowMagnificationPromptController
since this receiver is protected by a signature permission.

Bug: 161145287
Test: Manually verified logcat warning was not reported
Change-Id: I43b7fdd71b4a0427f8cd86b644325de289fbb259

Former-commit-id: 3cd0eb7e09ec0117b9d51d90388f8f389fe693a9