Remove DisplayContent#mExitingTokens and WindowToken#hasVisible
The mExitingTokens was only used for dumpsys, but almost never seen it
appear in log. The hasVisible is also removed by the way because it
was only used for checking whether to remove items in mExitingTokens.

Besides, WindowState prints mAnimatingExit and WindowContainer prints
animating animator in dumpsys. So it already provides enough info about
the exiting windows for debugging.

Bug: 163976519
Test: CtsWindowManagerDeviceTestCases

Change-Id: Ifaad007cb96e8f86a5a111f3dcf830ed25de3c20

Former-commit-id: fe493243d3d1e7226decd22077b7a665d741de31