Add apex-system-service tag to AndroidManifest.
This change introduces a new <apex-system-service> tag to the manifest.

A manifest may have multiple <apex-system-service> tags, with each tag
denoting one system service that lives in an apex. If the service is
not in SYSTEMSERVERCLASSPATH, the android:path attribute may be used to
indicate where the jar is located on the filesystem. The minSdkVersion
and maxSdkVersion attributes may be used to indicate which SDKs the
service should start on.

Test: atest PackageParserTest
Bug: 192880996
Change-Id: I28e595b397ed8db6c412f490f12c73537972efab

Former-commit-id: ad5d4baecbaee6e5c007d7d77beac0eac3339360