Enforce Coordinator encapsulation w/ Dagger
This allows for marking Coordinators with @CoordinatorScope.
Attempting to inject a Coordinator marked with this scope elsewhere in
SysUI will result in a build error.

This does *not* prevent marking Coordinators with @SysUISingleton,
which would allow for injection elsewhere.

Fixes: 204127880
Bug: 204468557
Test: mp droid
Merged-In: Ie79d632eb150790e6a9580b2c955af861bb2d182
Change-Id: Ie79d632eb150790e6a9580b2c955af861bb2d182

Former-commit-id: caeb5d278db81c82eff4d0bf669776d54a4f698e