Changed user check for play sound effect API
Changed the way the user is checked for audio manager play sound effect
API. Instead of capturing the process user, the current user is used
to determined if the audio should be played. Also moved the should
sound play check to audio service, as the check requires interact
across users permission, instead of adding permission to the API
this can be faciliated in the audio service.

Bug: 165038414
Test: atest AudioManagerTest#testSoundEffects
Test: atest AudioManagerTest#
	testCheckingZenModeBlockDoesNotRequireNotificationPolicyAccess
Test: atest VolumeDialogControllerImplTest
Test: manually disable audio effect sounds settings and play audio
NoNonSdkCheck: IAudioService#playSoundEffect is internal API
Change-Id: Ib0e3001fd76e3a5fdd831551d16f091029d03422

Former-commit-id: 21f2c4d59af6cd6dfc4e5ee1f2b11e058a05ebb9