Add config to detect e-call call based on dialed sim emerg # list.
We normally use the combined emergency number list from the emergency
number lists of both sims when placing emergency calls, with the provision
that the routing may be unspecified if a dialed number is an emergency
number on one sim but not the other.

This config allows a carrier to mandate that calls are only treated as
an emergency call on their network based on their own emergency number
list.

Test: TBD
Bug: 203947159
Change-Id: Id5cd4123f1bb7bdd85e564fc6997dfbc50652a5f
(cherry picked from commit 9430acf2f6f46b33dd7369dcad1a02243ab38642 [formerly 11dd6452d6621328b0bd430eeebacb55fd478cf3])

Former-commit-id: 43d9106a2c916ce8587b3b3b5006d3762511bf2d