Add public vibrate method with VibrationAttributes
Deprecate Vibrator.vibrate method that takes AudioAttributes and
introduce a new one that takes VibrationAttributes instead.

Update setAlwaysOnEffect to also take VibrationAttributes.

Introduce new public constructor for VibrationAttributes.Builder that
takes AudioAttributes, to be used in scenarios where the vibration
is meant to accompany an audio, and have similar attributes.

Bug: 163119475
Test: VibratorTest
Change-Id: I2721e3638aeb1aa4df3d515c3c0465923457dcc0

Former-commit-id: 4e9af660508dea185c011a6db78c1bc372e0f07e