Do not send split info update when TaskFragmentContainer
...in a split has no activities

If a TaskFragmentContainer in a SplitContainer has no
activities, then that means that either the entire split
is going to be removed or the empty TaskFragmentContainer
is about to get a running activity. This CL prevents a
split info update from sent in this case because the info
will soon be stale by another update.

Bug: b/204193051
Test: atest CtsWindowManagerJetpackTestCases:ActivityEmbeddingLaunchTests
Change-Id: I4b6f684b8d8e9a9b3fefd54b399d8a7d83c6cfb0

Former-commit-id: 5cb70ee9d46047cba342804d4f82936e4bbff516