BatteryService: remove HIDL-ness
Move HIDL-specifics of BatteryService to a list of classes
under com.android.server.health, so that we can replace
with the AIDL HAL easily.

HealthInfo types are kept as HIDL types. This will be
transformed to AIDL types in the next CL.

Also move BatteryServiceTest to HealthServiceWrapperTest,
since it essentially test just HealthServiceWrapper. Moving
the test to the proper package allows us to make things
package private.

Also make HealthServiceWrapperTest JUnit 4, since AndroidTestCase
is deprecated.

Test: builds
Test: atest FrameworksServicesTests
Bug: 177269435

Change-Id: Ie4dff341594fe9185a188ad963f29ade5d199e85

Former-commit-id: 8902c87ea99f1006ee96b0a5eb816011b1ca53c4