PiP: Recalculate size if too big/small on display change.
On display changes, it is possible that the old bounds is now too big
or too small for the new display. This is especially like on foldables,
such as a min size on a folded state is too big when in an unfolded
state.

Bug: 198679951
Test: Fold, minimal size, then unfold - see size change, and resizing
now doesn't jump to from too small -> small

Change-Id: I5de64860f1fbb476ff1a6a8dd9a0c50fd3259a45

Former-commit-id: 4ef153a7970c64f3cb997c408b41465e3e9ef36f