Set pending top and update process state for top resumed change
When switching focus between resumed activities, there is no activity
state change to trigger update process state. So if it is the case
that misses to update, add as pending top to ensure the current top
have higher oom-adj, and trigger a full oom-adj update to refresh
process states because layer rank of tasks are also changed.

Fixes: 204980211
Test: ActivityTaskSupervisorTests#testUpdatePendingTopForTopResumed
Test: Enter split with 2 apps, use command
      adb shell dumpsys activity oom | grep -A2 A/TOP
      to observe the value of oom-adj after switching focus.

Change-Id: I411a3980ea23d875dd8d9ce7af80de74957f1caa

Former-commit-id: 6a99e776867fa567a67416950fcd0b99e255e13f