Wrap animated dialog view inside another view
Before this CL, when animating a dialog from a View, we would extract
the original dialog Window background and set it as the background of
the original dialog content View. This would cause problems if that
original dialog content View had some paddings and if the background had
insets, as setting the View background would mess up the View paddings.

This CL now adds an additional View between the host dialog root and the
original dialog content Wiew. We set the original dialog Window
background to this additional View, so that the original dialog View
remains unchanged.

Bug: 193634619
Test: atest DialogLaunchAnimatorTest
Test: Manual, launch the User Switcher dialog and observe that the
      paddings are correct.
Change-Id: I64e688431c2d205f0cf200b9debfe195dc3e2364

Former-commit-id: f928f5b2e374e6b52466034ff251b23c882ca1a9