Remove all size compat UIs when the keyguard becomes occluded
This is needed because when a non resizable activity is launched over the lockscreen and a size compat restart button is displayed (after fold/unfold), clicking on it results in undefined behavior that seems broken. The solution to hide the restart button in such cases was a UX decision.

Fix: 202712269
Test: atest WMShellUnitTests:SizeCompatUILayoutTest
Test: atest SystemUITests:WMShellTest
Change-Id: Iced3948e0d7a37feb2f3422a7f25a22aab20a888

Former-commit-id: 6ed662a9db53fb9026ff62d6d351c1bc1ff25a57