Add user aware getPermittedInputMethods in DPM
In order to make on-screen keyboard settings
support tab layout. It's needed to get the
permitted input method list as work profile
user.

Add this @hide API in DevicePolicyManager which
should only be used in system.

This change only add the API first for further
usage, there is no user-facing changes.

Bug: 197707782
Test: atest com.android.server.devicepolicy.DevicePolicyManagerTest
      atest com.android.cts.devicepolicy.OrgOwnedProfileOwnerTest#testPermittedInputMethods
      atest com.android.cts.devicepolicy.OrgOwnedProfileOwnerTest#testPermittedInputMethodsLogged
Change-Id: I8ee9f523d2b25af7bfea2a138dba601f9678b8aa

Former-commit-id: 0f8da178c89344bea6153ed12142092812e589af