Add signature checking to AppCompatOverridesService.
As part of SLS allowlist implementation, security team has made
signature checking a requirement for allowlist. This change introduces
an optional string at the start of the flag which is the signature. If
the signature is present, it will be enforced.

The new format for the config string is:
<signature?>~<change-id>:<min-version-code?>:<max-version-code?>:<enabled>

Where signature, min-version-code, max-version-code are all optional

Test: locally on device
Test: atest FrameworksMockingServicesTests:AppCompatOverridesParserTest
Bug: 205279440
Change-Id: Ic78286ff315470ec1a0385231f5033d9c3339bbd

Former-commit-id: 6c4062561b3f98fff43f34f0150b3dbbe12aebff