Animate the Internet dialog height changes
This CL animates the height of the Internet dialog when it changes,
using a custom layout animator.

Note that I have tried using TransitionManager to get a better animation
(that would also properly fade elements in & out) but hit a few bugs
that made me lose confidence in TM (see b/203045010).

See b/201046726#comment20 for a video.

Bug: 201046726
Test: Manual
Change-Id: I02a0bfe67b705ea68e0d82a310a4361a94c1200d

Former-commit-id: b905ab10df84c9122d8c9f82232ce33516a9e3b0