Remove overlay when DC is mirroring
Layers can be added to the overlay when DisplayContent is mirroring
which would add content that's not from the original display. This
change puts the overlay layer offscreen so the root layer will only have
the mirrored content below.

This fixes an issue where screen recording with MediaProjection would
show a black frame as the first frame. That frame was the RotationLayer
getting put on top since the system thinks it needs to freeze the
display.

Test: No longer black frame as first frame in recording
Test: SurfaceViewSyncTest
Fixes: 205235272
Change-Id: I96323db05221d02db4d7762c7f36d3af904ef49e

Former-commit-id: 5ed18499acb99b992c9bd5f4e0f937695c4b632d