Account for the power allowlist.
Give apps on the power allowlist a higher minimum balance so they're
able to do more work. When a user removes an app from the list, we
reclaim some of the credits (with the amount reclaimed depending on the
time since the app was used).

Bug: 158300259
Bug: 202954395
Test: Charge device from 1% and see exempted apps get more credits.
Exempted apps approach minimum balance as device nears 100%.

Change-Id: Id19c43d31477daa90173bda5294479487b025ea2

Former-commit-id: 2a25c41813a67f835c760b4d11cd325068211da7