Multiuser perfetto metric: adb tracepoints
Add tracepoints in adb for user switches and creation, for the
primary purpose of performance metrics in trace_processor.

These need to be done at the level of shell, not
UserController/UserManager because each entrypoint needs an explicit
tracepoint.
We already have some for UI-based entrypoints - and we want to time
from the moment
the user requests the switch, not some internal point that might be
substantially later - so to capture adb-based switching/creation within
the same trace processor mechanism, we must also add tracepoints for
the shell commands.

Bug: 192001071
Test: manual
Change-Id: I2d7d30faf7906eb455f5d2dc05f0d6b2d96518c2

Former-commit-id: 12d2a30a3e944615090f01c8be4c65cba8bdf8dd