Turn on TV when pressing power or wake key during boot
Applies to HDMI-CEC playback devices only.

Boot isn't always triggered by the user and therefore shouldn't turn on an HDMI-connected TV.

However, if a power or wake key is pressed during boot, there is a clear user intention to use the Android device.
This CL makes the playback device turn on the connected TV in that
situation.

Test: atest PowerManagerServiceTest and atest com.android.server.hdmi
Bug: 183691245
Change-Id: I53e54d72a5c7fa54d27d743d80bad4644e897d2b

Former-commit-id: 736d5f0d8132bf67466b13c2e881050606c21a4f