Un-deprecate TelephonyManager SIM access APIs
The APIs was deprecated starting from R release and recommended to
switch to OMAPI which has better resource management. However, OMAPI
does not provide the equevalent functions and has no plan to
support them in the near future.

With no resource to unify the two sets of public APIs, the plan
is to continue investing the current TelephonyManager SIM
acccess APIs. With target of maximum backward compactibility
and minimum architecture change, the plan here focuses on
improving reliability (with better resource management) and
usability (with better documentation).

As the first CL of the feature, it only un-deprecates the APIs.

Bug: 197658986
Test: make
Change-Id: If23dcfa4dda45efd91d1c8b6970691390d7243c8

Former-commit-id: 9fda707679bddd381afb83bcf3646a09caebc12d