[DO NOT MERGE] Make extra navigation bar count towards non decor frame
The extra navigation bar should be treated in the same way as the
navigation bar to avoid compatibility issue. There are devices only
contain an alternative extra navigation bar but not the default
navigation bar. The app may receive conflict information from the insets
system and the window configuration in that case.

Test: CtsWindowManagerDeviceTestCases (on device with task bar)
Bug: 205496695
Change-Id: Idef3d09c723ffcd67372db29fc0cc51a59cc0aa1

Former-commit-id: b356de42b082a8057ed5f25c6c9ad172b963e72d