Animate the power menu when triggered from QS
This CL animates the power menu when it is triggered from its QS button.
It does that by first adding support for fullscreen dialogs in
DialogLaunchAnimator, given that the power menu dialog is fullscreen for
legacy reasons.
Note that I first tried to migrate it to a non fullscreen dialog, but I
hit issues with the ConstraintLayout + Flow that are used to display the
grid of the power menu items (see b/206097842 for more info).

In this CL, I also removed:
 - Custom window flags that were added in the past and don't seem to be
   needed anymore.
 - The handwritten animation when triggering from the power button. We
   now simply fade the dialog in/out like we animate all the other
   dialogs (approved by UX in b/202257171#comment16). I will improve
   this default animation in a follow-up CL too.
 - Some dialog reinflation that would happen when the configuration or
   UI mode would change. For the former, I still make sure that we
   recompute the maximum number of items per row. For the latter, it
   would not happen while this dialog is shown as changing the UI mode
   would require dismissing the dialog first anyways.

See b/202257171#comment11 for a video.

Bug: 202257171
Test: Manual
Test: atest GlobalActionsDialogLiteTest
Change-Id: Ib5facb579c8dce114f29c07502a4a10b50d281be

Former-commit-id: 95b7f9b0719bd2eb04913e74b5a47b9db8fdda64