Correct exported state for InputMethodManagerService receiver
The InputMethodManagerService receiver that listens for
CLOSE_SYSTEM_DIALOGS broadcasts was initially marked as not exported
since this broadcast should be dropped by the platform when sent from
apps. However since this broadcast is still valid to be used in tests,
receivers registering for this broadcast should be marked as exported.
This commit updates the receiver to exported to allow any tests that
send this broadcast to complete successfully.

Bug: 205796926
Test: atest ImeInsetsVisibilityTest
Change-Id: I92cb5cab839699084b3330c88fda7d72c78c565b

Former-commit-id: c79ea6623c818707f7bf91da124b2cdb97cbdb32