Sort remote cast notifications to the end
Remote cast notifications are posted by gmscore when the screen turns
on. Since they are not user-initiated, and since lastUpdated won't be
meaningful, sort them behind any other active media controls in the
carousel.

To avoid hard coding packages, we determine whether the notification is an
RCN by whether it is from a privileged app + sets a substitute app name,
which is only done by RCN currently - as noted in the comment we plan to
add an explicit API for this in the T release

Fixes: 204935142
Test: atest com.android.systemui.media
Test: manual
Change-Id: Ia5d1e1f544b546644941f05efda199844aae74e5

Former-commit-id: d0a83fe892edf2149868910c54026643e0226848