Fix regression when launching the media output dialog
This CL fixes a regression that was caused by ag/16191211: when opening
the media output dialog, we would animate from the media dialog content
instead of its parent (that has the dialog background).

This CL changes that by detecting that the view we are launching from is
actually the content of a dialog that was animated using
DialogLaunchAnimator, and animates its parent to also animate the
background.

See b/200781817#comment8 for before/after videos.

Bug: 200781817
Test: Manual
Change-Id: I20cf8bdff8b6d29e9aa458f96d65851a0298f98d

Former-commit-id: 79bd0cafdd2234da9c9cf1fd21b4f81c8ee381a6