Release surface control of remote transition leash without removal
Because the transition-leash is put under transition root, it will
be removed with the root when the finish transaction executes.

Otherwise if "transition-leash" is removed earlier than reparenting
task to its original parent, it may show a few frames with empty root
and looks flickering.
 Transition Root
  > transition-leash
     > Task
  > transition-leash
     > Task

So it is enough to only release the reference.

Bug: 206062800
Test: adb shell setprop  persist.debug.shell_transit 1; reboot
      Launch app from home and return to home. The app and home
      should not flicker.
Change-Id: I451f4bcce90405a6256ec40afd345824a79c96ae

Former-commit-id: c9853d9a63b1013f35c2441d3c2b99fb78c7b940