[DO NOT MERGE] Animate dialog stack and use in UserSwitcher
Allows to dismiss a dialog stack (dialogs launched from dialog content)
by dismissing the top one to the original touchSurface. After this, all
the intermediate dialogs are dismissed.

Use this in the user switcher dialog.

When adding users, the dialog that opens to set the name and picture is
not part of this stack (it's a separate activity), so the dialog stack
is collapsed before launching the activity.

Test: atest SystemUITests
Bug: 185360864
Change-Id: Ife8e0f9189783b69034ce660d1075c3f76fbbf06

Former-commit-id: 327d67cb2a2604b85e1b3e0cb72f957471db88ba