Support row and column title for the accessibility node
The accessibility service, Talkback, wants to support more information
about the table on Chrome, but there’s no sources for it. We need to add
more data in the Accessibility node to make Talkback can do what it
wants.

This patch is adding the row and column title in the inner class,
CollectionItemInfo, of the AccessibilityNodeInfo class.

Bug: 181789503
Test: a11y CTS & unit tests
Change-Id: Id5d267836acf92b10a33a11677bd8f3ed040bfb9

Former-commit-id: 4b3593c0f71e613a0687a93fd671919af0e724e6