Try to fix ActivityThreadTest failure on foldables
The root cause of this failure is that we don't specify Configuration#orientation
in test, so #handleActivityConfigurationChanged uses device orientation
instead, which is landscape in the failure case. In this way, TestActivity receives
2 onConfigurationChanged callbacks, one is to make Activity portrait,
the other is to make Activity landscape, which is from process
configuration.

This CL specifies orientation in each propagated Configuration to
prevent orientation overridden by process Configuration.

Bug: 204903019
Test: atest ActivityThreadTest
Change-Id: I4d7720ef1128fc8612381b1a076b06beb765f1b6

Former-commit-id: 74cc4db965deeb2e16a455deeec1b30669b709db