Better support for Resource based flags.
Introduce ResourceFlag class and ParcelableFlag classes, from which
other flag types are based.

ResourceFlags always pull there default values out of
context.getResources. They can not, however, be directly
serialized and sent to the flag app.

For that, we now have ParcelableFlag. Prior to this change, we
were not properly alerting the Flag app about the current value of
flags when it asked. This change now also fixes that.

All flags are coerced into ParcelableFlag before being sent to the
app.

Note that ResourceFlags perform runtime logic, even in release builds,
meaing that the opportunity for optimizing them at compile time is
much smaller. Tools such as proguard will not have the opportunity
to strip them out entirely.

Bug: 203548827
Test: atest SystemUITests
Change-Id: I3c37ccae16c00237f9cffffc9c22580454ecb8cc

Former-commit-id: 62ac77294a73a3395f49eab3691c1e1094a1baae