Remeasure entire carousel view when players change
The page indicator view was not being consistently remeasured when new dots
were added for new players, resulting in invisible (0x0) dot views

Instead of only requiring remeasure of the mediaCarousel view, require
it for the entire frame which contains both that and the pageIndicator
view. This ensures that the dots will be measured and sized correctly.

Fixes: 194351796
Test: manual - follow repro steps, observe correct number of dots
Change-Id: I2b2b8b5bc906549f76d2705e117d0a2670da819d

Former-commit-id: 8ea639575b4fb4389e0560879982c0c9ed9422af