Privapp IntentResolver += PACKAGE_USAGE_STATS
This permission is required to query AppPredictionService.

When the unbundled Chooser doesn't request this permission,
we get a SecurityException attempting to invoke
createPredictionSession(). When we request the permission
but don't add it to this allow-list, the device boot-loops.
With the permission added in both places (and some other
local changes), the unbundled Chooser is able to show the
full set of app predictions.

Test: Manual, as described above.
Bug: 206831012
Change-Id: I154d0555471f7f8820ac151b77a61cb8d9bf7603

Former-commit-id: 42fb76f068aab7c92d6f2d1cb08898357c506966