Support subscription for CarrierService#onLoadConfig
CarrierService#onLoadConfig is called from platform to carrier service
to load carrier-specific configuration values.

In case when the multiple subscriptions are supported for the same
carrier (e.g. Fi on pSIM + Fi on eSIM case), the current API is not
able to distinguish on which subscription the carrier config should
be loaded. The subscription parameter introduced in the API can
resolve the problem.

If carrier app has no need to distinguish configs by subscriptions,
it can simply ignore overriding the new version of the API.

Bug: 131697791
Test: atest CarrierServiceTest
CTS-Coverage-Bug: 210143342
Change-Id: Id7cd75c24f01119749cb5bf31d4335c638547aaf

Former-commit-id: bbec7b6a5f3bcab8cc9138103b9a76ebcdd8a2cb