Fix GL memory not being removed from lost RAM
Devices upgrading to S are not required to update their memtrack HAL
implementation and won't support reporting the total GPU private memory.

In these cases it's likely the HAL reports 0 for the total GPU private
memory, since in S the HAL defition only overloads the already existing
getMemory() API to report this.

Return an error value in these cases to prevent swapping the sum of GL
type memory for all processes with the total GPU private memory when not
supported as doing this effectively leaves the GL memory as part of
lost RAM.

Bug: 203245390
Test: dumpsys meminfo
Change-Id: I478143e435b047bf349bd4ec6808707a5db1505b

Former-commit-id: 0b7c1774ba42daef7c80bf2f00fe1c0327e756ae