Introduce BackupTransportClient
This class wraps the calls from the framework to the remote
IBackupTransport service. The framework will now interact with this
wrapper instead of calling IBackupTransport directly. This allows us to
shield the client code (when needed) from changes in IBackupTransport API as
well as introduce additional call management (e.g. cancell calls to the
remote transport early if needed).

Bug: 202716271
Change-Id: Idf6c5918d631537382d56ca053923ec663c88b48

Former-commit-id: 78a5c42baf9bc458b913759b5757f2b4374f2379