Adds trace messages when goAsync and finish are called.
These will be used to troubleshoot ANRs. In this case calling goAsync
and forgetting to call finish on the PendingResult could be the cause.
We have existing instrumentation to include these in Perfetto traces
and have used the "am" tag to avoid having to allowlist every single
app of interest in the Perfetto tracing configuration (the "am" tag is
allowlisted).

Test: Manual: captured a Perfetto trace and verified the counters are
present.
Bug: b/207823767

Change-Id: I96aadf4d6d891f3c7d140b9792e7f893243d9943

Former-commit-id: 9b47623790d9ee71973a4c502d4851a36efcd04f