Block split Activity in lock task mode
Before this CL, Activity was only blocked from finishing if it is the
last Activity in th Task to start lock task mode.
However, it may make half of screen black if the finishing Activity is
in one of adjacent TaskFragments.
This CL considers the scenario that two adjacent TaskFragments side-by-side.
If a finishing Activity leads to TaskFragment empty, and also make the
adjacent TaskFragment finish, block the Activity from finishing.

Bug: 205813709
Test: manual - use Settings in task pinning mode

Change-Id: I6116d644aeccc13460d302fe3b40953fa67dca6c

Former-commit-id: 0399d9c9514d7f1b350d2206dd24dff075a8da89