KeyguardRotateController handles minimum show time
Instead of KeyguardIndicationTextView ensuring that
keyguard messages are shown for a minimum amount of time,
move logic into KeygaurdIndicationRotateTextViewController,
since the rotating text view controller already handles
the rotating text durations its easier to coordinate
the timing in one class rather than across two.

Update the charging info if the device is at full charged and
is plugged in. Sometimes the battery status is not updated
so we need  to check the actual %.

Test: manual, atest KeyguardIndicationControllerTest
KeyguardIndicationRotateTextViewControllerTest
Fixes: 188134239
Fixes: 206342960
Fixes: 195347563
Change-Id: Id0cf9bb6c5ac95f83dbb2dcf32f41fa719c0853e

Change-Id: If67b1b1cef2fd6db554df105005a185b1ad0f7d2

Former-commit-id: 69af3501822f30be26deece49243cae3fbb34419