Add VirtualDevice.launchPendingIntent
Add launchPendingIntent method on VirtualDevice, which allows specifying
a display ID to launch the pending intent on. For activity intents, this
is equivalent to adding the ActivityOptions with setLaunchDisplayId on
the pending intent. For service and broadcast pending intents,
VirtualDeviceManager will try to catch any trampolines activities
launched by the pending intent creator and start those on the virtual
display instead.

Bug: 194949534
Test: Manual, CTS to be added in ag/15781910
Change-Id: Ia5d6c92167f0ee6c9f246fd4335daa824aceaef9

Former-commit-id: f03d832c434cc327f7dbc317324c26e4cd1b1b68