Un-@Ignore 7 passing ChooserActivity tests.
Of 69 ChooserActivityTest cases, 22 are currently
ignored, but should be restored in order to maximize
coverage in preparation for the new unbundled
implementation (but note that these tests are still
only run against the "bundled" system chooser for now).

Of the 22 ignored tests, 7 of them pass reliably and can
be un-ignored immediately (although it's not known whether
they were ignored for a reason other than failures/flakes).
The remaining 15 will require some changes to the test
and/or ChooserActivity code, to be investigated & completed
in future CLs.

Bug: 208803748
Test: `atest --iterations 100 ChooserActivityTest` (100
iterations ensures no regressions-to-flakiness.)

Change-Id: Iede5835b429cccfd65fdc114397460fc00781a75

Former-commit-id: ec29931867f47c50e462b718f0cf45649f5a986d