DO NOT MERGE Listen for config changes when updating features.
Listen for configuration changes to push out folding feature changes.
Fixes a bug where the reference implementation would not emit the
correct folding feature if a configuration changed. The effect was that
you would see the folding feature for portait while in landscape or
vice-versa.

Maintains parity with Sidecar providing updates for Activities that do
not handle configuration changes.

Bug: 205342008, 206697963
Test: Manual - Open the sample app and use the application that &&
  handles configuration changes.  &&
  Expect the folding feature to match the physical orientation
Test: Manual - open an app that handles PiP and folding features &&
  Flattend the device and put app into PiP &&
  Fold device and put PiP app in full screen. &&
  Expect the app to react to fold.
Test: atest CtsWindowManagerJetpackTestCases
Change-Id: I726827c9f1482ddd0b55a6839e34660deb340767

Former-commit-id: 6c39d4fb3c851b024498c271d5660ac773d4f2fc