Fix flaky assertion
Sometimes, when quick switching between apps, the previous app is restarted (e.g., it was stopped by the OS to release memory). Because of this, the app launch creates a SnapshotStartingWindow before the secondary app is visible again. So we now check if this is visible if the Window itself isn't yet.

Test: atest FlickerTests SwitchImeWindowsFromGestureNavTest#imeAppWindowVisibility

Bug: 208628307
Change-Id: I1cc48da5f1dd9ad78b8565a0028272dd3b985a39

Former-commit-id: cd355907712d0e0aa5e392d9b55834b9a27486a9