Increasing size of lock icon on big screen devices
Lock icon can be very small on big devices because by default its padding is specified in pixels (48px) and its size is calculated in LockIconViewController to be 64dp.
When density is small enough that doesn't leave enough space for normal size of lock icon.
Now padding is overriden for sw600 to be density independent.

Fixes: 208440478
Test: manual on two different large screen devices on master and sc-v2
Change-Id: I73e1e06c23d782951b9cc45fb7bee349a2405593

Former-commit-id: 2bcb8113908fe5429a8ce57303f7021b57b421d2