Auto-deny requests after some user denials
After the user denies the request to add a tile (a particular
ComponentName) a certain number of times, auto deny further requests for
that user,component pair. Cancelling the dialog (by going back/click
outside) don't count against.

This information is reset when the package is uninstalled or data is
cleared for the package.

In the future, this information will be persisted across restarts.

Fixes: 208671346
Test: manual
Test: atest com.android.server.statusbar

Change-Id: I1828d1d3e462fc9b4af94c4350df895994d02b6c

Former-commit-id: 3a738b180eb939a52617876b85785f2f278d4b87