use test app's data folder as filter directory
During a retry, test methods to be retried are written into a
file, pushed onto device and read by the test instrumentation
for execution.

The default location for such filter file is /data/local/tmp,
which is prohibited for access by system uid packages, such
as the system UI tests package.

This change uses the test app's data folder for filter: this
is not generally applicable for all the tests since root
permission is required, but it is a relatively safe assumption
in case of testing against system uid pakcages since it
implies userdebug build is being used.

Bug: 152001555
Fixes: 208473429
Test: http://go/forrest-run/L76200000952166377
Change-Id: Ifc268a1cdb94ff8e47b6bd323464deb28e0c20ea

Former-commit-id: 7d6a7c5f0e3ff342674312e48717b1fe8f2d1e58