Pass MTE RuntimeFlags to AppZygote.
Fix AppZygote process and its children (the actual service processes)
ignoring android:memtagMode attribute in the app manifest.

When starting a new AppZygote process, apply memtag-related flags as
determined by the app manifest (and modified by compat features and
actual h/w capabilities). If this is not done, MTE is always disabled in
an AppZygote process, which makes it impossible to enable it in the
AppZygote's children.

This change has no effect unless MTE is supported in the hardware
(ARMv9) and enabled in the system.

Bug: 207557677
Test: CtsTaggingHostTestCases
Change-Id: Ibf64db8882a1fbffe6c0cc3cd3bc2299b088876a

Former-commit-id: 2e50afbd53c686a7cb82c73e926ab34d7dd9257d