Reset the settings key of the migration tooltips as enabled.
Root cause:
The users don't know how to remove the new floating menu on the
home screen when they upgrade from R to S.

Goals:
Show the tooltip for the users to guide them to settings page when they
upgrade to the S_MR version. They
could adjust the mode configuration as “gesture mode” to solve it based on the feature b/190563948.

Bug: 207354119
Test: manual test
Merged-In: I1c877af7658781c32cddb2bb819013f1037409de
Change-Id: Ia3685dfa36e8f571b2d0268af2616aecb56fc90b

Former-commit-id: ec99fb3cb399311b79e5aea5517a8315b4fc1f63