Fixes re-setting activity type when embed activity in split-screen
An exception was thrown by setting the activity's type to
TYPE_UNDEFINED while the activity was already set as TYPE_STANDARD.
It happens while embedding an activity (which reparents the activity
into a TaskFragment) in split-screen.

Activity is no longer be the direct child of a leaf Task while the
activity being embedded. So, we should use WC#getActivity() vs.
checking if the task has any child while updating the activity type.

Bug: 208803173
Test: Drag settings from taskbar into split
Change-Id: Ia2e3dc251ed1908d0b088794853760c4f511138e

Former-commit-id: 6deec52fa975d3cdc6b5ed4a15eb3bb74b092f5c