Reset the settings key of the migration tooltips as enabled.
Root cause:
The users don't know how to remove the new floating menu on the
home screen when they upgrade from R to S.

Goals:
Show the tooltip for the users to guide them to settings page when they
upgrade to the S_MR version. They
could adjust the mode configuration as “gesture mode” to solve it based on the feature b/190563948.

Bug: 207354119
Test: manual test
Change-Id: I1c877af7658781c32cddb2bb819013f1037409de

Former-commit-id: 053a3e313a37fa703c40cb761b593261f608dee0