Move out allowlists for com.android.cellbroadcast
This change separates out the allowlist for the
com.android.cellbroadcast mainline module. This apex bundles two apks
(CellBroadcastApp and CellBroadcastServiceModule) that have regular
apk counterparts for non-mainline devices.
The allowlists are now in the projects which contain the source code for
the respective APKs, and are marked as dependencies for them

Test: boots
Bug: 190375768
Change-Id: I98c7c5a0ffb9b16dd859cb586d3afae082981d7c

Former-commit-id: 8f5e78c0a773c9d4ef245797d884b5cb37e47054