Add theme information during throwing exception
When an application defines themes that are not inherited from
android:style/Theme, it may cause UnsupportedOperationException.
The cause is that the attributes needed by the application are
defined in neither theme nor layout attribute set.

It takes really a long time to trace all of the themes applied.
This patch logs the basic information about themes to help the
developers to triage issues.

Test: manual test described in b/209708201#comment3

Bug: 209708201
Change-Id: I440aa18d9627adf5e35fd74d7adb1c895f0b169d

Former-commit-id: dcd3a9cbe94d4a1cebd3e87e7a35ad2e47073fd9