CommunalSource.Connection Introduction
This changelist introduces the Connection interface,
abstracting details around the result reporting mechanism.
It also sets the expectation that the client holds onto
the connection, allowing connectors to build around this
assumption.

The Connection interface also allows connection state
to be contained in a single entity, tied to the primer.
As a result, the CommunalSources no longer need to
track state and maintain their own callbacks.

Test: atest CommunalSourcePrimerTest
Bug: 209607168

Change-Id: I599d1ba421d76ec9e9c20c82718104358929c5e4

Former-commit-id: bb905fd00949f6b7747b5f5261b704a6962e1862