Do not override animation when the parent bounds doesn't support split
Without this cl, we always override animation as long as there is any
organzied TaskFragment. And users may see different animation on the
same small screen before and after they have unfold and fold.

The idea is to only register the remote handler if the app currently has
at least one split rule that can fit the current Task bounds.

When there can be split, we want to always override the animation to
make it consistent.

When there won't be any split, we want to allow app to override
animation, so they can still provide the same experience as normal phone
on small screen.

Fix: 207804116
Fix: 196173550
Test: manual test with fold/unfold
Change-Id: I3ae194f89b35afc5734338fd3bb350e0cd364c85

Former-commit-id: c0acf094f24d772f4e69d0fc5e0db62a8c900d4a