Fixing displaying two carriers in split shade header
For dual carriers all carrier information will be displayed inside QSCarrierGroupController and carrier information in StatusIconContainer should be hidden.
The difference in split shade is that for single carrier name of the carrier will be also displayed on the left side in QSCarrierGroupController.

Test: SplitShadeHeaderControllerTest and too much mocking
Fixes: 194178072
Merged-In: Idd2819fbe3d8b2b23f89999052cb900698ac4b11
Change-Id: Idd2819fbe3d8b2b23f89999052cb900698ac4b11

Former-commit-id: 12763375e5fae5572230c0be5a1616251a76e06a