Change SBKVM to use notification shade window instead of status bar
It's not clear to me that this code was strictly incorrect, but the
former `StatusBarKeyguardViewManager#getRootViewImpl()` method
referenced the StatusBar view. This code was written at a time when the
notification shade (and thus keyguard) was hosted in that window.

However with the change to split out NotificationShade into its own
window, this is no longer the host view for KeyguardViewMediator to care
about. This change replaces the use of StatusBar#getViewRootImpl to
instead use the NotificationShade window's viewRootImpl.

Also remove unused fields from SBKVM

Test: atest SystemUITests
Bug: 209005990
Change-Id: I752fd920cce8dc2be1ef7f3f6644741f3a0899fb

Former-commit-id: 3822f89d643380f709679eede198cece64caba75