Reset inherit-parent bounds in finishTransaction
Since shell doesn't know about inherited bounds, reset
them in finishTransaction so that tasks aren't left with
an outdated crop.

This is mainly to remedy an issue with rotation where the
animation (in shell) needs to set the task crop for the
duration of the animation, but doesn't know if it needs to
unset it or leave it.

Bug: 210065870
Test: open app, rotate to landscape, dismiss app to home,
      rotate to portrait, launch (resume) app.
Change-Id: I8fcfbe6d444cc3a5387be216e5ea9a1d4fb12a7e

Former-commit-id: 2dd6381527758345693b52107d05eb629291ea8e