Centering big clock vertically on large screen lockscreen and AOD
Big clock was not centered vertically properly in case we don't support smartspace on the device. When smartspace is not present big clock's height is wrap_content and that makes it (and all parent views) takes much less space. Then even though clock might be centered in parent, parent is not centered on the screen so it doesn't change much.
This change is making it always go match_parent, that is (more or less) full height of the screen.

Fixes: 209013448
Test: manual
Change-Id: I721b5750ae34d204c6eb0bdb7edca6aa12e3a88b

Former-commit-id: af3f020b35075184294e90cb5f1da6545081612b