[pm] move mPm related code from ScanPackageHelper to InstallPackageHelper
Tried to consolidate some of scanPackageNew and scanSystemPackage but
there is some delicate difference that I am not 100% certain to merge.
In this CL I moved all the code that requires mLock into
InstallPackageHelper, and made ScanPackageHelper a utility class. I will
keep doing the same for other helper classes, such as
DeletePackageHelper.

BUG: 199937291
Test: manual
Change-Id: I9f93cc36494e1fff84628026949ca5887bb3e323

Former-commit-id: 6321aedb64abc78992cd30858ee9434a107a7101