Update dropzones to match the size of the existing split
The dropzones were always split exactly in half on the screen
which looks weird if you adjust the split to a different size
and then drag new apps into split.

This adjusts the dropzones to match the size of the existing
split.

It also fixes a bug where the existing split bounds weren't
being properly calculated in the policy so the drag and drop
highlight wouldn't switch to the correct side until you
passed through the middle instead of where the divider
actually is.

Bug: 209504662
Test: manual - 1) have apps in split
               2) adjust the divider
               3) start dragging a new app into split, pass
                  across the divider
               => the highlight should immediately switch
               => the highlight should match the size of the
                  existing split
               4) test in landscape & portrait
Change-Id: I3c7a288625ad766d0c467d713a451fe8bfc1d669

Former-commit-id: c57aa4a84ea596fedee8d9e12962a06607325a30