Don't reset the keyguard if it's going away.
If we re-lock the keyguard *while* it's unlocking, we end up in a state where the AOD
UI is visible, but we are not actually locked and can unlock the device with no
authentication. This is possible with and without remote animations enabled, but is
easier to trigger when they're enabled.

"Keyguard reset" is done if we go to sleep while the keyguard is showing, and causes
the keyguard views to be reset. This is not sufficient if the keyguard is going away
as it should be re-shown instead of only reset.

Fixes: 210705299
Test: atest SystemUITests
Test: interrupt unlock by locking again, verify the device is actually locked
Change-Id: Ie7aeab50cdd8fb6e3a79f6ef1e0cb21ca61d665d

Former-commit-id: 757317e2e2aef1284c8ebbbe8d81d45c88fcc307