Fix issue where surface was animated while wake and unlocking.
This resulted in the wallpaper being briefly visible during wake and unlock, as the
surface alpha/translation animated. This is not the desired behavior, since we are
waking up from black/AOD, we don't want the surface to animate. The light reveal
scrim will make it visible.

Fixes: 201659992
Test: manual/atest SystemUITests
Merged-In: Id72f3124ee6c2c5c0867ac3cea8c2b04daa208c8
Change-Id: Id72f3124ee6c2c5c0867ac3cea8c2b04daa208c8

Former-commit-id: f853c0f15fa856dd28e6fe8801617c932484c40d