Keep DeviceFeatures updated in local devices.
Local devices currently contain an instance of HdmiDeviceInfo.
With the addition of DeviceFeatures to it, it should be kept updated
whenever a device's set of supported features is computed.

This also necessitates updating the builder method for <Report Features>
messages to use the new DeviceFeatures type.

Note: Updates to a local device's HdmiDeviceInfo are not propagated to
HdmiCecNetwork (except immediately after address allocation)
because the convention for the more dynamic fields of HdmiDeviceInfo
is to only update them in the local device.

Bug: 204854610
Test: atest com.android.server.hdmi
Change-Id: I1ba6f1bdfb36b36f079ec03063d5a6ce6d355e92

Former-commit-id: abbbf2ae456b518a581406bf54b2097c00ad4b8d