Give ActivityRecordInputSink its own surface
LetterboxSurface uses a z value of -1 and since it is a direct child
of the ActivityRecords SurfaceControl it get placed below the
ActivityRecords surface. This caused AcitivtyRecordInputSink to receive
touches that previously went to the LetterboxSurface. To solve this
ActivityRecordInputSink is assigned its own surface that is set as a
child of the AcitivtyRecord. By using Integer.MIN_VALUE for the Z value
it is ensured that it is behind all Activity related windows and only
consume touches that would otherwise pass through.

Test: Manually verified that that layers are correctly ordered (with
winscope and adb shell dumpsys input)
Bug: 194480991
Bug: 210593865

Change-Id: Ib4122948f19c144c455343e53f29af0e70197c00

Former-commit-id: 2771c9eac5717ba1984cb9dadf1cc3efe3f679a4