Fix CTS failure for waterfall device
When entering to or exiting from Pip mode, the PipTaskOrganizer will
set the windowing mode of the activity of the task to
WINDOWING_MODE_FULLSCREEN and temporarily set the bounds of the task to
fullscreen size for transitioning.
It will get the wrong smallestScreenWidthDp if the calculation is based
on this temporary fullscreen bounds.

We should just inherit the value from parent for this temporary state.

Bug: 210892943
Test: atest PinnedStackTests
Change-Id: Iaf83de8ad13768c755df7b2a32edbbbec8a02317
Merged-In: If3001b13b48b1e3124b9bd898ab740a2d3025df0

Former-commit-id: 667db9ef20f23f8e484587025788a07efd204faf