Fix wallpaper zoom effect with shell transition
Currently WallpaperController#adjustWallpaperWindows is called for
shell transition at DisplayContent#ensureActivitiesVisible and
Transition#onTransactionReady for collecting visibility change of
wallpaper. But if wallpaper target is cleared and wallpaper is
requested to be invisible, its offset and scale won't be updated.

So this change allows visible window to update wallpaper during
transition. And make sure the offsets are calculated based on the
configuration that reported to client, to avoid mixing coordinates
with different orientation.

Bug: 211447882
Test: adb shell setprop persist.debug.shell_transit 1; reboot
      Set static image as wallpaper (don't use live wallpaper).
      Launch portrait and landscape app from portrait home.
      Observe the wallpaper should show the same zoom effect.
Change-Id: I1db9207c79bc45dedd90e7b13af5517f09abe4a7

Former-commit-id: 5eeb4070c1a432b5a599554adf30c76afe30c002