Clean some unnecessary abstraction of ATMI about keyguard
Since ATMS and WM were put in the same package with the same lock,
the abstraction is no longer needed and can be called directly.
Move notifyKeyguardFlagsChanged to DisplayContent because most of
the operations are associated to the display.

This contains one logic change is that if isKeyguardLocked() is false,
then skip notifyKeyguardFlagsChanged() because visibility won't change.

That reduces unnecessary invocation of root.ensureActivitiesVisible
such as an activity uses FLAG_SHOW_WHEN_LOCKED while device is unlocked.

Bug: 163976519
Test: CtsActivityManagerDeviceTestCases

Change-Id: I13e7d965df62c44eb07e9db7c7ac33ab990212bd

Former-commit-id: 0dc092cc02d2c7dc69a15a0bde2dbeb9682811aa