Makes it so changes to device state that have the FLAG_CANCEL_OVERRIDE_REQUESTS flag cancels all override requests
Previously if a device state with the FLAG_CANCEL_STICKY_REQUESTS flag
was entered, it only cancelled requests that were made from a process
that was no longer alive. This was causing undesirable behavior where
a device state that should act as a reset state was not cancelling
override requests. This change makes it so if the device enters a
state that is defined as a reset state (has the
FLAG_CANCEL_OVERRIDE_REQUESTS on it) all the override requests are
cancelled. This change also changes the flag name to
FLAG_CANCEL_OVERRIDE_REQUESTS to be more correct and descriptive.

Bug: 207686851
Test: DeviceStateTest, DeviceStateProviderImplTest
Change-Id: I0af1d2917a3d422b5d05f0210d627454c23f9636

Former-commit-id: 34ac32953cf9df8c439bf4e318ff09247ccac853