Add GameSession APIs for starting activities.
This API allows the game service provider to start an
activity in the existing game task and receive a result
through the GameSessionActivityCallback instead of the
result being sent to the game.

For an example use case, in the Android S dashboard, the
dashboard starts the YouTube-hosted live stream activity to
handle YouTube live streaming and the Play Games-hosted
dashboard activity to handle showing Play Games services
data.

The API allows the GameSession provider to respond to
different activity results returned from the target
activity.

Test: manual e2e testing
Bug: 208242220
Bug: 202414447
Bug: 202417255
CTS-Coverage-Bug: 206128693
Change-Id: I1eb017ade1ae60a713d85a70df23845498206df7

Former-commit-id: ed33816c43e373fdbc6adc07bdeaf9d7394c4246