Fixed an issue where the user could be stuck with pulse expanding
Added a lot more logging and dumping to the LS shade controller
and the pulse expansion handler.
Also added a safeguard and fixed a bug where the pulse expansion
wasn't reset in lockedDown mode.
Fixed an issue where aborting the unlock in docked down shade wouldn't
reset the dragdown amount.
Always allowing the set the dragdown amount if it's a reset.

Fixes: 209839390
Fixes: 198098523
Test: see repo steps in bug, avoid it manually, run regular tests
Change-Id: I3398763141e8f828c00fb5d5184d64ffc66d3237

Former-commit-id: a7617fc89d406b400da0e827969f71911fc9a9cc