Run sound trigger HAL watchdog on uptime clock
Previously, it was implemented based on java.util.Timer, which uses a
wall clock. That is problematic, because that clock doesn't stop when
the system goes to sleep, and so if that happened in the middle of
executing soundtrigger HAL code, the watchdog would fire as soon as
the system woke up.

This change implements the watchdog using an uptime-clock-based timer.

Fixes: 199798343
Test: atest FrameworksServicesTests:com.android.server.soundtrigger_middleware.UptimeTimerTest
Merged-In: I913336837aff3e2c74c77f32aac24348612803f9
Change-Id: I5ce6ba997a7af616785634223af8c6139d030111

Former-commit-id: 21a296bf479bb300465b5be2fdba1968afd51a51