Fix IME layering target above IME surface
CL[1] fixed the IME flickering issue when switching recents tasks with
specifying to set IME relative layer until the layering target is not
animated and be IME control target.

However, it didn't consider the case that the IME layering target could
be the drop down box window that not same as the IME control target,
so that the issue happens when we didn't set IME above this window,

the app drop down box will be above IME and cosumed the input to hide
the keyboard.

Fix this regression by checking the token of the layering target to set
IME relative layer if the target isn't transitioning and same as the
token of the control target.

[1]: Iae65d6047c02dae241f2df21c666a03cca58fd0b

Fix: 209722479
Test: manual as test steps
Test: atest ZOrderingTests#\
        testAssignWindowLayers_ForImeOnPopupImeLayeringTarget

Change-Id: Id2e700c20b964e296576c49391853b680418bf57

Former-commit-id: 096919c66454f7592737a8f77f3b518b966f6925