Add a benchmark test for Resources.getIdentifier
In order to prevent unstable benchmark results because of cache,
a random resource name with an expected value comes from
Random.nextInt(10001).The range is from 0 to 10000.

This patch creates 10001 color resources with the prefix
"i_am_color_%x".  The range is from "i_am_color_0" to
"i_am_color_2710". The benchmark measures randomly accessing any
of those resources.

Test: atest --request-upload-result \
    CorePerfTests:android.app.ResourcesPerfTest#getIdentifier

Bug: 191487225
Change-Id: I357b43fda77e31b6d234233fd902eb8e4696301e

Former-commit-id: ccb8c015b7fe8c782b4013047bdd7d61cb8d11cf