Avoid reflection in ObjectPrinter
This works around the fact that the Android toolchain may now increase
the visibility of fields, so a field being public is not a good signal
for whether it should be printed.

Replaced object printing with toString() for composite objects,
relying on the auto-generated printer for AIDL objects.

In the process, cleaned-up a bunch of unused methods / arguments and
lint warnings.


Bug: 210901706
Test: Manual verification of STMW logs and dumpsys
Change-Id: I839e3830e9378ecf8fca9d4d0d3fdcffe00c0c6c

Former-commit-id: e788a3e16ad86ea9439a8c5c0187f270b2e69490