Make sure that animation leash is removed with window container
Unlike ActivityRecord and WindowState, the general window container
doesn't clean up its animation for removal. So if the container is
removed while animating, the animation leash surface may be leaked.

One case is that when rotating display with a short life window such
as Toast, the the leash created by FadeRotationAnimationController
may remain on DisplayArea (the parent of WindowToken).

Bug: 212483446
Test: atest WindowContainerTests#testRemoveImmediatelyClearsLeash
Change-Id: I1b4833b14a3c6c4ea1338156f5bfd7bf475645d1

Former-commit-id: 19c7260b395467a22e0c883d8d1bc929ce697d91