Deprecate SoftInputWindowProto.{name,window_type,gravity,takes_focus}
This is a follow up CL to my previous CL [1], which made it clear that
the following 4 tracing fields in SoftInputWindowProto had always been
logged with the same values.

  - name:         "InputMethod"
  - window_type:  WindowManager.LayoutParams.TYPE_INPUT_METHOD
  - gravity:      Gravity.BOTTOM
  - takes_focus:  false

This CL simply deprecates those fields in favor of simplicity and
smaller code size.

There should be no behavior change unless IME tracing is enabled.

 [1]: I619ae6011f5e66dba126931719569a97513eaa7e
      f8ecc8fac393c0c7eb090f294a7d17a3829009ce [formerly ea7801bd267f98fba10741b88bc900aef9f84121]

Bug: 192412909
Test: IME tracing still works
Change-Id: I04b59307ecf2c692282c7882ac839f77c94b938f

Former-commit-id: ef936ff07dcf97bb5f6ccabc411367aaeb0cdb06