Support following typing focus in window mode [3/n].
There are 3 milestones in this feature.
1. Refactor the callbacks for Accessibility in WindowManagerInternal.
2. Implement this feature in such new architecture.
3. Implement the setting choice in preference page.

This CL is for the 3rd milestone.

We add a ACCESSIBILITY_MAGNIFICATION_FOLLOW_TYPING_ENABLED key to store
a user preference for following typing focus. The default value of
preference is true.

If the value is false, we wouldn't do anything in
onRectangleOnScreenRequested and just return.

If the value is true, we would dispatch the event to an activated
magnification controller.

Bug: 194668976
Test: atest AccessibilityManagerServiceTest
	atest AccessibilityUserStateTest
	atest FullScreenMagnificationControllerTest
	atest MagnificationControllerTest
	atest WindowMagnificationManagerTest
	atest SettingsBackupTest
Change-Id: I7203698d2edcc474ab1638ac81e72e054a61d6fd

Former-commit-id: f4e9ca2c79749d215af5c89680ad153b1259a80c