Fix Keyboard not hiding when dismissing remote input from shade
In RemoteInputView will invoke hideSoftInputFromWindow when the
editor view is invisble. However, the timing will be after the editor
view lost focus, so that in hideSoftInputFromWindow will ignore the
request since the editor view has been detached.

As the notification shade is still focused, fix it by calling
getWindowController().hide(ime()) when the IME is visible.

Fix: 209925343
Test: manul as steps:
   1) Make/install EditTextVariations
   2) Launch EditTextVariations app, select "Direct-Reply" from menu.
   3) after direct-reply notificaion popup,
      scroll down the notification panel.
   4) Tap "Direct Reply Test" button on the notification to show IME.
   5) Swipe out direct-reply notification to remove
   6) Expect IME will be hidden

Change-Id: I6739d3bd6ad7dbaa4c0d430c57d1167644cf8584

Former-commit-id: d5d6db78d243c7fe49e715120f99784041bfa976