Add Jank test for unfold transition
A new TransitionProgressListener is added to start and stop the jank monitor during the transition. This is a no-op if the jank monitor is disabled.

The reason why it is added in StatusBarWindowController is that InteractionJankMonitor needs an attached to begin, and the status bar should always be present during the animation.

Bug: 213862831
Test: Running test suite in forrest with config in g3
Change-Id: I99dcd4e879b9c2c47611096b29b164881e8714c7

Former-commit-id: caf468dce4755e376d02aaab86457ac2b5f8e2e0