Include app badge without markings in bubble data
Sometimes we need to show the app badge without any markings for
priority conversations. For example in case of the manage menu.
As we include the priority ring in the app badge icon itself, add
another app badge icon to bubble data that does not include the priority
ring - raw app badge.

Change-Id: Ib472519d9bec5cdd09d7559e4dad2ee5e5f71b50
Test: verified that priority conversation bubble has ring around app
badge and no ring in manage menu
Bug: 189173831

Change-Id: I276e110ece40bf394f58da9acc3d0ecd09cf5afc

Former-commit-id: 4311ec42bfaa40993cd1112026f7b0f4a23bd0e8