WindowManager Task: Update surface visibility
Currently task never updates it's Surface visibility
and it relies on its children becoming invisible. This has been ok
as activity visibility was always managed correctly. As we move towards
attaching overlays to tasks though, we need to fix this issue so that
overlays will automatically be made invisible when the task is
dismissed.

Bug: 213603716
Test: Existing tests pass
Change-Id: I5806d69d34cfbad7717c405852e025634f3947a2

Former-commit-id: 84a606287e52d874e588899629933f9debbe450d