Fix an issue where the divider bar wasn't part of the hitRect for split
This doesn't result in an issue with existing code but with some changes
in how the animation works, this will become an issue.

e.g. you can get a null target when it should be non-null which changes
whether the animation is a switch between targets vs a new drag into a
target.

Test: manual - with new animation code, put two apps in split and then
               drag a new one into split and move it between containers,
               the animation should look like a switch (color changes
               seamlessly rather than a jump to transparent).
Bug: 207011867
Change-Id: I6d9c892654cbd57742bdf3268a8ab343f0fd3d3b

Former-commit-id: 5650fa8ee47061d4a24276bad3f65ab796c72e4d