Replace some netd calls in NetworkManagementService
Some netd binder calls which used in NetworkManagementService are
deprecated, replace them with new ConnectivityManager APIs. The only
caller for those replaced functions are NetworkPolicyManagementService.

Although NetworkPolicyManagerService starts before ConnectivityService,
it is fine for replacement because the order in SystemServer is
create ConnectivivtyService -> NPMS#bindConnectivityManager ->
NPMS#systemReady. NPMS#initService is called under NPMS#systemReady, it
guarantees that ConnectivityManager will always be initialized when it
is called.

Bug: 209935649
Test: m
Change-Id: I2142f997ae1086d1cc3f71d5b2f5c9cfe0466a76

Former-commit-id: 86b957648bf1ef7524e73cbe00e6d20d35d0198d