Change path for DefaultSelectionToolbarRenderService
We met the ClassNotFoundException when putting the service under
system service folder. Cannot find the root cause to fix it now. We
need to more time to find the reason so the workaround is to move the
class to the framework/base/core/service/selectiontoolbar to fix the
problem now. We will go back to fix the problem later.

Bug: 214122495
Test: manual. Can bind to DefaultSelectionToolbarRenderService and
the methods can be called successfully.

Change-Id: I9218c9ceee25a105fe592c0c63905ba2e2f8a305

Former-commit-id: 895098daeba27fca060b55924f302265438b0fe5