Fixed silent mode alarm play at full volume
During do not disturb mode, privilege alarms sometimes were not
played at full volume. This was happening when alarm was disabled
and the setting to play priviledge alarms at full volume was enabled.
Specifically, the new event for device changed was decrementing the
privilege alarm active count which was incremented during the playback
start event. The device change event was resetting the count to zero,
signaling that there were no more privilege active alarms and thus
reducing the volume back to zero when it should have played at full
volume.

Bug: 211324401
Test: 1) Set volume to minimum for alarms
      2) Enable do not disturb mode
      3) Turn off alarm in do not disturb
      4) Send ETWS text message
      5) Alarm should be heard at full volume
Change-Id: Id591975ef6bf1b72ab9a4d23e98c93603e0ab914

Former-commit-id: de364ea6476c060a6449b842ef306c328d38929e