[Media TTT] Don't use futures for determining whether a transfer has succeeded or not.
After discussing with davidsamuelson@, we decided that using futures to
communicate across processes was error-prone. Specifically, on SysUI
we'll switch to a failure state if we don't get a response from the
future after a certain amount of time. However, it's possible that the
transfer that D2DI kicks off ends up succeeding after we show the
failure chip, so the user would get incorrect information.

To avoid this, we're just going to have D2DI notify us when the transfer
starts and when it succeeds or fails.

Bug: 203800643
Bug: 203800347
Test: verify `adb shell cmd statusbar media-ttt-chip-add-sender Tablet
TransferInitiated` stays loading forever and never switches to succeed
state.
Test: media.taptotransfer tests

Change-Id: I38521e9af463a2d8eafdea63f9cf7c32bbbea9c6

Former-commit-id: 4d4b6c47caeed71f8764f77e8c5fb3e1071944da