Allow alertOverride if the summary GROUP_ALERT_ALL
Because we still require all children to have GROUP_ALERT_PARENT, the meaning of GROUP_ALERT_ALL is equivalent to GROUP_ALERT_PARENT in any case where we would have an alertOverride, so we now allow either of those values.
Also of note, because the alertOverride absolutely needs to have GROUP_ALERT_PARENT, and the semantics of heterogeneous child alert behaviors is unclear, we're keeping the requirement that all children have GROUP_APERT_PARENT to avoid over-applying this logic.

Bug: 185680162
Test: atest NotificationGroupManagerLegacyTest
Change-Id: I63a51cc1d94fc8a8617ca76a0fa479e3829816c6

Former-commit-id: 6ee64f6f263f183b27f7b24d30d74bce1ceef660