Make InputMethodManagerService#MSG_* private if possible
Message formats used in InputMethodManagerService#handleMessage() are
supposed to be internal details of InputMethodManagerService.
Allowing other classes to directly post messages to
InputMethodManagerService would make it difficult for us to track the
source code.

This CL makes InputMethodManagerService#MSG_* as private constants as
many as possible, which should have no behavior changes.

Bug: 192412909
Test: presubmit
Change-Id: Ifdde313c884b9ec10e45fa6449e7541217f5fb24

Former-commit-id: 0467d70d13dc9897edffb510b1cf301b7fcc79bc